Fix ActiveAdmin JS
While working on the mess that is #157, I discovered that the JS part of our ActiveAdmin setup was broken. This commit does not add a test, but this will be tested in an upcoming system spec that only passes with working ActiveAdmin JavaScript. This is extracted from the jumble of #157.