Allow user to change title on PeopleSoft side
Prior to this patch, the ETDs controller did a Submission.find_or_initialize_by with both the dissertation ID and the title. We do not need the title to find a dissertation, though, and this thwarts finding the correct ETD when a user changes the title. Instead, separate this call into a find_by (with diss ID), or a new (with diss ID and title), which is logically identical to what find_or_initialize_by does.